Created a new whitelist file interpreter.
This interpreter now supports the Process Type Tag (<Whitelisted>, <Games> etc), which can sort the loaded processes into more specific categories such as Whitelisted and Games.
